ParamEditor: Change from process to thread for Windows compatibility

+++ MAVProxy/modules/mavproxy_paramedit/param_editor.py
@@ -5,6 +5,7 @@ Akshath Singhal
 June 2019
 '''
 
+import platform
 from MAVProxy.modules.lib import mp_util
 from MAVProxy.modules.lib import multiproc
 from MAVProxy.modules.mavproxy_paramedit import ph_event
@@ -96,11 +97,19 @@ class ParamEditorMain(object):
         self.mpstate.param_editor = self
         self.needs_unloading = False
 
-        self.child = multiproc.Process(
-                        target=self.child_task,
-                        args=(self.event_queue,
-                              self.event_queue_lock, self.gui_event_queue,
-                              self.gui_event_queue_lock, self.close_window))
+        if platform.system() == 'Windows':
+            self.child = threading.Thread(
+                            target=self.child_task,
+                            args=(self.event_queue,
+                                  self.event_queue_lock, self.gui_event_queue,
+                                  self.gui_event_queue_lock, self.close_window))
+        else:
+            self.child = multiproc.Process(
+                                target=self.child_task,
+                                args=(self.event_queue,
+                                      self.event_queue_lock, self.gui_event_queue,
+                                      self.gui_event_queue_lock, self.close_window))
+
         self.child.start()
 
         self.event_thread = ParamEditorEventThread(
@@ -205,7 +214,10 @@ class ParamEditorMain(object):
         '''close the Parameter Editor window'''
         self.time_to_quit = True
         self.close_window.release()
-        self.child.terminate()
+        if platform.system() == 'Windows':
+            self.child.join()
+        else:
+            self.child.terminate()
 
     def set_params(self):
         for param, value in self.paramchanged.items():
